CG tests pass
The Fisher-information matrix is always singular, so we can't just select a random b and compute F^-1*b. Luckily, the gradient is in the column space of F, since anything in the nullspace of F doesn't contribute to the loss at all.

TODO: prove that conjugate gradients will actually solve a system, even if F is singular, given that there is a local maximum.