use native numerics for environments
Ideally, environments shouldn't care about the agent. By forcing Envs to use anyvec, we were compelling them to deal with agent-specific information (e.g. the numeric type, or GPU support, etc.). This was not only theoretically incorrect: it was also extremely slow in certain cases (e.g. when doing a lot of GPU vector operations that could easily be done on the CPU).